+++ scratch.js
@@ -3,7 +3,12 @@ var nlp = require('./src/index');
 
 //(private|general|major)
 
-console.log(nlp('Björk, the “singer-songwriter”').normalize().out('text'));
+// console.log(nlp('the so-called ❛singer-songwriter❜').normalize().out('text'));
+let doc = nlp(`"seven"`);
+// console.log(doc.list[0].terms[0]);
+console.log(doc.values().toNumber().out('text'));
+
+// console.log(nlp('Director of the F.B.I').acronyms().addPeriods().out('text'));
 
 // nlp('about this ...').normalize().debug()
 // console.log(nlp('word ...').toLowerCase().out());