+++ test/unit/result/normalize.test.js
@@ -28,7 +28,11 @@ test('normalize():', function(t) {
   [
     [' so... you like DONUTS? have all the donuts in the WORLD!!!', 'so you like donuts? have all the donuts in the world!'],
     ['This is a test. .', 'this is a test.'],
-    ['Björk, the singer-songwriter', 'bjork the singer songwriter']
+    ['Björk, the singer-songwriter...', 'bjork the singer songwriter'],
+    ['the so-called “fascist  dictator”', 'the so called "fascist dictator"'],
+    // ['the so-called ❛singer-songwriter❜', 'the so called \'singer songwriter\''],
+    // ['the so-called ❛group of seven❜', 'the so called \'group of 7\''],
+    ['Director of the F.B.I.', 'director of the fbi'],
   ].forEach(function(a) {
     var str = nlp(a[0]).normalize().out('text');
     str_test(str, a[0], a[1], t);